# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for sports field booking system - Take Home Test",
        "title": "Sports Field Booking API",
        "contact": {
            "name": "API Support",
            "email": "support@sagaratech.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3005",
    "basePath": "/api",
    "paths": {
...
        "/payments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Process mock payment for a booking. **AUTHENTICATED USERS** can pay for their own bookings. This is a mock implementation for testing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Process payment (Mock)",
                "parameters": [
                    {
                        "description": "Payment data including booking_id and payment_method",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/take-home-test_app_models.PaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input or booking not found",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User cannot pay for this booking",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            }
        },
        "/payments/notification": {
            "post": {
                "description": "Webhook endpoint for Midtrans payment notifications. **NO AUTHENTICATION REQUIRED** - This endpoint is called by Midtrans server to update payment status. Do not call this endpoint directly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Handle payment notification webhook",
                "parameters": [
                    {
                        "description": "Midtrans notification payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            }
        },
        "/payments/{booking_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get payment details for a specific booking. **USERS** can access payments for their own bookings. **ADMINS** can access all payments.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment by booking ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID (UUID format)",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/take-home-test_app_models.PaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden - User cannot access this payment",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Payment for specified booking does not exist",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            }
        },
        "/payments/{booking_id}/transaction": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Midtrans payment transaction for a booking. Returns payment URL for redirection to Midtrans payment page. **AUTHENTICATED USERS** can create transactions for their own bookings.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create real payment transaction (Midtrans)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID (UUID format)",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/take-home-test_app_models.PaymentTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid booking or payment error",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User cannot create payment for this booking",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get authenticated user's profile information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user details by ID (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "take-home-test_app_models.BasicResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {},
                "pagination": {
                    "$ref": "#/definitions/take-home-test_app_models.Pagination"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "take-home-test_app_models.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "booking_id",
                "payment_method"
            ],
            "properties": {
                "booking_id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                }
            }
        },
...
        "take-home-test_app_models.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "take-home-test_app_models.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "booking_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "take-home-test_app_models.PaymentTransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "payment_id": {
                    "type": "string"
                },
                "redirect_url": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}